requirements.txt
python-telegram-bot==21.4
python-dotenv
requests
aiosqlite>=0.17
